+++ src/main/java/neresources/api/distributions/DistributionSquare.java
@@ -10,7 +10,7 @@ public class DistributionSquare extends DistributionBase
      * @param maxY last occurrence
      * @param chance chance for the distribution
      */
-    public DistributionSquare(int minY, int maxY, double chance)
+    public DistributionSquare(int minY, int maxY, float chance)
     {
         super(DistributionHelpers.getSquareDistribution(Math.max(minY, 0), Math.min(maxY, 255), chance));
         this.bestHeight = (minY + maxY) / 2;
@@ -24,9 +24,9 @@ public class DistributionSquare extends DistributionBase
      * @param max0 end of ramp down
      * @param chance the chance at the top
      */
-    public DistributionSquare(int min0, int minY, int maxY, int max0, double chance)
+    public DistributionSquare(int min0, int minY, int maxY, int max0, float chance)
     {
         super(DistributionHelpers.getRoundedSquareDistribution(min0, minY, maxY, max0, chance));
-        this.bestHeight = DistributionHelpers.calculateMeanLevel(this.getDistribution(),(minY+maxY)/2,0,100D);
+        this.bestHeight = DistributionHelpers.calculateMeanLevel(this.getDistribution(),(minY+maxY)/2,0,100F);
     }
 }
