+++ src/main/java/neresources/api/utils/DistributionHelpers.java
@@ -8,16 +8,16 @@ public class DistributionHelpers
      * @param midY the top, middle of the triangle
      * @param range length of the sides
      * @param maxChance chance at the top
-     * @return an array of 256 doubles in triangular distribution
+     * @return an array of 256 floats in triangular distribution
      */
-    public static double[] getTriangularDistribution(int midY, int range, double maxChance)
+    public static float[] getTriangularDistribution(int midY, int range, float maxChance)
     {
-        double[] triangle = new double[range * 2 + 1];
-        double modChance = maxChance / (range+1);
+        float[] triangle = new float[range * 2 + 1];
+        float modChance = maxChance / (range+1);
         for (int i = 0; i <= range; i++)
             for (int j = 0; j <= range; j++)
                 triangle[i + j] += modChance;
-        double[] result = new double[256];
+        float[] result = new float[256];
         for (int i = 0; i < triangle.length; i++)
         {
             int mapToPos = i + midY - range;
@@ -32,11 +32,11 @@ public class DistributionHelpers
      * @param minY first occurrence
      * @param maxY last occurrence
      * @param chance the chance
-     * @return an array of 256 doubles in square distribution
+     * @return an array of 256 floats in square distribution
      */
-    public static double[] getSquareDistribution(int minY, int maxY, double chance)
+    public static float[] getSquareDistribution(int minY, int maxY, float chance)
     {
-        double[] result = new double[256];
+        float[] result = new float[256];
         for (int i = minY; i <= maxY; i++)
             result[i] = chance;
         return result;
@@ -48,20 +48,20 @@ public class DistributionHelpers
      * @param maxY start of the ramp down
      * @param max0 end of ramp down
      * @param chance the chance at the top
-     * @return an array of 256 doubles in square distribution
+     * @return an array of 256 floats in square distribution
      */
-    public static double[] getRoundedSquareDistribution(int min0, int minY, int maxY, int max0, double chance)
+    public static float[] getRoundedSquareDistribution(int min0, int minY, int maxY, int max0, float chance)
     {
-        double[] result = new double[256];
+        float[] result = new float[256];
         addDistribution(result, getRampDistribution(min0, minY, chance), min0);
         addDistribution(result, getSquareDistribution(minY, maxY, chance));
         addDistribution(result, getRampDistribution(max0, maxY, chance), maxY);
         return result;
     }
 
-    public static double[] getUnderwaterDistribution(double chance)
+    public static float[] getUnderwaterDistribution(float chance)
     {
-        double[] result = getTriangularDistribution(47,8,chance/7);
+        float[] result = getTriangularDistribution(47,8,chance/7);
         addDistribution(result,getRampDistribution(57,62,chance),57);
         result[62] = chance;
         addDistribution(result,getTriangularDistribution(55,4,chance/3));
@@ -72,18 +72,18 @@ public class DistributionHelpers
      * @param minY first occurrence
      * @param maxY last occurrence
      * @param maxChance chance at the top of the ramp
-     * @return an array of doubles with length |maxY - minY| in ramp distribution
+     * @return an array of floats with length |maxY - minY| in ramp distribution
      */
-    public static double[] getRampDistribution(int minY, int maxY, double maxChance)
+    public static float[] getRampDistribution(int minY, int maxY, float maxChance)
     {
-        if (minY == maxY) return new double[0];
+        if (minY == maxY) return new float[0];
         if (minY > maxY) return reverse(getRampDistribution(maxY, minY, maxChance));
 
         int range = maxY - minY;
-        double[] result = new double[range + 1];
+        float[] result = new float[range + 1];
         for (int i = 0; i < range; i++)
         {
-            result[i] = (maxChance * (double) i) / range;
+            result[i] = (maxChance * (float) i) / range;
         }
         return result;
     }
@@ -93,7 +93,7 @@ public class DistributionHelpers
      * @param add the to add distribution
      * @return the sum of both distributions
      */
-    public static double[] addDistribution(double[] base, double[] add)
+    public static float[] addDistribution(float[] base, float[] add)
     {
         return addDistribution(base, add, 0);
     }
@@ -104,7 +104,7 @@ public class DistributionHelpers
      * @param offset the first element from the base array to start adding to
      * @return the sum of both distributions
      */
-    public static double[] addDistribution(double[] base, double[] add, int offset)
+    public static float[] addDistribution(float[] base, float[] add, int offset)
     {
         int addCount = 0;
         for (int i = offset; i < Math.min(base.length, add.length + offset); i++)
@@ -116,9 +116,9 @@ public class DistributionHelpers
      * @param array
      * @return a reversed version of the given array
      */
-    public static double[] reverse(double[] array)
+    public static float[] reverse(float[] array)
     {
-        double[] result = new double[array.length];
+        float[] result = new float[array.length];
         for (int i = 0; i < array.length; i++)
         {
             result[array.length - 1 - i] = array[i];
@@ -133,15 +133,15 @@ public class DistributionHelpers
      * @param difference the difference
      * @return the mean level of the distribution
      */
-    public static int calculateMeanLevel(double[] distribution, int mid, int oldMid, double difference)
+    public static int calculateMeanLevel(float[] distribution, int mid, int oldMid, float difference)
     {
-        double totalUp = 0;
-        double totalDown = 0;
+        float totalUp = 0;
+        float totalDown = 0;
         for (int i = 0; i < distribution.length; i++){
             if (i<mid) totalDown+=distribution[i];
             else totalUp+=distribution[i];
         }
-        double newDifference = totalUp-totalDown;
+        float newDifference = totalUp-totalDown;
         if (Math.abs(difference+newDifference)<=Math.abs(newDifference)) {
             if (Math.abs(newDifference) < Math.abs(difference))
                 return mid;
@@ -156,9 +156,9 @@ public class DistributionHelpers
      * @param num the denominator
      * @return the divided array
      */
-    public static double[] divideArray(double[] array, double num)
+    public static float[] divideArray(float[] array, float num)
     {
-        double[] result = new double[array.length];
+        float[] result = new float[array.length];
         for (int i = 0; i < array.length; i++)
             result[i]=array[i]/num;
         return result;
@@ -169,26 +169,26 @@ public class DistributionHelpers
      * @param num the multiplier
      * @return the divided array
      */
-    public static double[] multiplyArray(double[] array, double num)
+    public static float[] multiplyArray(float[] array, float num)
     {
-        double[] result = new double[array.length];
+        float[] result = new float[array.length];
         for (int i = 0; i < array.length; i++)
             result[i]=array[i]*num;
         return result;
     }
 
-    public static double[] maxJoinArray(double[] array1, double[] array2)
+    public static float[] maxJoinArray(float[] array1, float[] array2)
     {
-        double[] result = new double[array1.length];
+        float[] result = new float[array1.length];
         if (array1.length != array2.length) return result;
         for (int i = 0; i < array1.length; i++)
             result[i]=Math.max(array1[i],array2[i]);
         return result;
     }
 
-    public static double sum(double[] distribution) {
-        double result = 0;
-        for (double val:distribution)
+    public static float sum(float[] distribution) {
+        float result = 0;
+        for (float val:distribution)
             result+=val;
         return result;
     }
